requirements.txt
autopep8==1.5.3
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
pycodestyle==2.6.0
python-dotenv==0.13.0
selenium==3.141.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.9
